+++ src/importexport/import/internal/au3/au3importer.h
@@ -23,7 +23,9 @@ public:
     Au3Importer() = default;
 
     void init() override;
+
     FileInfo fileInfo(const muse::io::path_t& filePath) override;
+
     bool import(const muse::io::path_t& filePath) override;
     bool importIntoTrack(const muse::io::path_t& filePath, trackedit::TrackId dstTrackId, muse::secs_t startTime) override;
