+++ src/projectscene/CMakeLists.txt
@@ -159,6 +159,8 @@ set(MODULE_SRC
     ${CMAKE_CURRENT_LIST_DIR}/view/tracksitemsview/splittoolcontroller.h
     ${CMAKE_CURRENT_LIST_DIR}/view/tracksitemsview/mousehelper.cpp
     ${CMAKE_CURRENT_LIST_DIR}/view/tracksitemsview/mousehelper.h
+    ${CMAKE_CURRENT_LIST_DIR}/view/tracksitemsview/dropcontroller.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/view/tracksitemsview/dropcontroller.h
 
     ${CMAKE_CURRENT_LIST_DIR}/view/timeline/timelinecontext.cpp
     ${CMAKE_CURRENT_LIST_DIR}/view/timeline/timelinecontext.h
